requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
openai>=1.12.0
pydantic>=2.0.0
pymupdf>=1.23.0
pandas>=2.0.0
python-dotenv>=1.0.0
hydra-core>=1.3.0